requirements.txt
arabic-reshaper==2.1.3
asgiref==3.4.1
autopep8==1.6.0
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.2.8
django-crispy-forms==1.13.0
django-filter==21.1
django-xhtml2pdf==0.0.4
future==0.18.2
gunicorn==20.1.0
html5lib==1.1
Pillow==8.4.0
psycopg2==2.9.2
psycopg2-binary==2.9.2
pycodestyle==2.8.0
PyPDF2==1.26.0
python-bidi==0.4.2
pytz==2021.3
reportlab==3.6.2
six==1.16.0
sqlparse==0.4.2
static3==0.7.0
toml==0.10.2
webencodings==0.5.1
whitenoise==5.3.0
xhtml2pdf==0.2.5
